+++ src/main/java/com/novelbot/api/dto/auth/LoginRequest.java
@@ -6,7 +6,7 @@ import lombok.*;
 @Setter
 @NoArgsConstructor
 @AllArgsConstructor
-public class LoginAuthDTO {
+public class LoginRequest {
     private String username;
     private String password;
 }
